consolidate into one file and clean up
